+++ src/supported-documents/index.js
@@ -9,7 +9,24 @@ export type CountryType = {
   name: string,
 }
 
-export const getSupportedCountriesForDocument = (documentType) => {
+export const getCountryDataForDocumentType = (
+  countryCode: string,
+  documentType: string
+) => {
+  // Consistent with API, which accepts a 3-letter ISO country code for issuing_country param value
+  if (countryCode && countryCode.length === 3) {
+    const supportedCountriesList = getSupportedCountriesForDocument(
+      documentType
+    )
+    const country = supportedCountriesList.find(
+      (countryData) => countryData.country_alpha3 === countryCode
+    )
+    return country
+  }
+  return null
+}
+
+export const getSupportedCountriesForDocument = (documentType: string) => {
   switch (documentType) {
     case 'driving_licence':
       return getCountriesList(supportedDrivingLicences)
@@ -23,7 +40,7 @@ export const getSupportedCountriesForDocument = (documentType) => {
   }
 }
 
-const getCountriesList = (supportedDocsData) => {
+const getCountriesList = (supportedDocsData: Array<object>) => {
   const countriesList = supportedDocsData.map((docData) => {
     const { sourceData } = docData
     return {